refactor: extract types from shared to be types
