+++ src/app/modules/submission/submission.service.ts
@@ -15,7 +15,6 @@ import {
   AttachmentPresignedPostDataMapType,
   AttachmentSizeMapType,
   FormResponseMode,
-  ParsedClearAttachmentFieldResponseV3,
   SubmissionMetadata,
   SubmissionMetadataList,
   SubmissionPaymentDto,
@@ -29,6 +28,7 @@ import {
   SubmissionData,
 } from '../../../types'
 import {
+  ParsedClearAttachmentFieldResponseV3,
   ParsedClearAttachmentResponse,
   ParsedClearFormFieldResponse,
 } from '../../../types/api'
